requirements.txt
flask
Flask-SQLAlchemy
Flask-Migrate
python-dotenv
mysqlclient
sqlalchemy_serializer
flask-marshmallow
marshmallow-sqlalchemy